Refactor ride service to use events for sending emails
