+++ BACKEND/src/main/java/com/project/backend/events/RideCreatedEvent.java
@@ -1,0 +1,11 @@
+package com.project.backend.events;
+
+import com.project.backend.models.Ride;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class RideCreatedEvent {
+    private Ride ride;
+}
